pyproject.toml
[tool.poetry]
name = "gps-stuff"
version = "0.1.1"
description = ""
authors = ["Emiliano Jordan <emiliano.jordan@protonmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Flask = "^2.1.3"
alembic = {extras = ["tz"], version = "^1.8.1"}
pandas = "^1.4.3"
matplotlib = "^3.5.2"
dash = "^2.6.0"
SQLAlchemy = "^1.4.39"
psycopg2-binary = "^2.9.3"
requests = "^2.28.1"
Flask-Migrate = "^3.1.0"
Flask-SQLAlchemy = "^2.5.1"
python-dotenv = "^0.20.0"
flask-shell-ipython = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipython = "^8.4.0"
pytest-alembic = "^0.8.3"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
)/
'''

[tool.isort]
profile = "black"
